solved TODO of duplicated code
