+++ lib/src/span_scanner.dart
@@ -29,13 +29,17 @@ class SpanScanner extends StringScanner implements LineScanner {
 
   @override
   set state(LineScannerState state) {
-    if (state is! _SpanScannerState || !identical(state._scanner, this)) {
+    if (!isClassState(state)) {
       throw ArgumentError('The given LineScannerState was not returned by '
           'this LineScanner.');
     }
 
     position = state.position;
   }
+
+  @override
+  bool isClassState(LineScannerState state) =>
+      state is _SpanScannerState && identical(state._scanner, this);
 
   /// The [FileSpan] for [lastMatch].
   ///
